requirements.txt
django==3.2.0
pytest==7.1.1
pytest-cov
djangorestframework==3.13.1
pytest-django==3.7.0